extract load-data worker
