+++ modules/gob-worker-load-data/source/__tests__/remove-duplicate-areas.test.js
@@ -1,9 +1,9 @@
 /* eslint-env jest */
 // @flow
 
-jest.mock('../../../db')
+jest.mock('@gob/web-database')
 
-import db from '../../../db'
+import db from '@gob/web-database'
 import removeDuplicateAreas, {
     generateOps,
     buildRemoveAreaOps,
