+++ modules/gob-worker-load-data/source/__tests__/store-data.test.js
@@ -1,9 +1,9 @@
 /* eslint-env jest */
 // @flow
 
-jest.mock('../../../db')
+jest.mock('@gob/web-database')
 
-import db from '../../../db'
+import db from '@gob/web-database'
 import storeData, {storeArea, storeCourses} from '../store-data'
 
 import {mockArea} from './area.support'
