+++ modules/gob-worker-load-data/source/get-cache-store-name.js
