+++ modules/gob-worker-load-data/source/lib-prepare-course.js
