+++ modules/gob-worker-load-data/source/load-files.js
