+++ modules/gob-worker-load-data/source/parse-data.js
