+++ modules/gob-worker-load-data/source/remove-duplicate-areas.js
@@ -7,7 +7,7 @@ import fromPairs from 'lodash/fromPairs'
 import sortBy from 'lodash/sortBy'
 import map from 'lodash/map'
 
-import db from '../../db'
+import db from '@gob/web-database'
 
 type AreaOfStudy = {
     name: string,
